Add a check.cpp and main README.md
Added a README.md file that explaines how to build the project with c++
make utility.
check.cpp displays the distribution of one of the files (in this example,
the one with normal distribution). It shows the min and max of the numbers
and draws a graph.
Also modified Makefile to build and execute check.cpp.